+++ CHANGELOG.md
@@ -3,8 +3,9 @@ base on EasyStream.jl
 学习
 
 v0.2.0 2021.8.15
-1. table 的 stream. 
-
+1. 修改变量名, 加注释. 
+2. 加入op到op_1.jl中
+3. 扩展 connectors
 
 
 
@@ -18,9 +19,17 @@ TODO
 
 -------------------------------------------------------------------------
 重设计:
+参考 Transducers.jl, OnlineStats.jl, DataTools.jl, chain.jl, streamz, flink 因为要兼容. 
+
+应该是 每个op 有自定义的state. stream上也可以有内置的state 
+
 stream
 
-op(stat,data) -> nothing
+op(stat,data) -> stat,data
+数据和状态都要 在ops 之间传输 
+
+
+
 
 
 代码生成: op_1(stat), 会自己生成 一个 struct对象.